fix: exclude yv8 models

+++ src/daemon/misc.py
@@ -67,9 +67,9 @@ async def upload_files_to_s3(bucket: str, local_path: str, s3_path: str, suffixe
             err(f"Could not find {local_path}")
             return 0
 
-        # Recursively glob all files in the local path
+        # Recursively glob all files in the local path, excluding anything with 'yolov8' in the path
         for obj in pathlib.Path(local_path).rglob('*'):
-            if obj.is_file():
+            if obj.is_file() and 'yolov8' not in obj.parts:
                 for s in suffixes:
                     if obj.suffix == s:
                         if await upload_file(obj, bucket, f'{s3_path}/{obj.name}'):
